Avoid parsing "[::1]" as having scheme "["
We've slightly modified the URI_MATCHER regular expression components
to avoid parsing the plain string "[::1]" as having a scheme of "[" and
host of ":1]". That said, Junk is passed in and Junk was returned so,
it's not exactly the most ideal thing, it just makes our lives a little
less awful since we can avoid it.