Guard against wide-domain checks
Previously, we would parse the malicious URL example with an authority
of 'user:pass@xdavidhu.me\\test.corp.google.com:8080' but we would not
parse that into its components because it had invalid characters. So
accessing the `host` attribute would result in `None`. That said,
someone might still have used the `authority` attribute and been misled.
To avoid misuse by developers, let's parse this similarly to the fix in
the blog post.

See also:
- https://bugs.xdavidhu.me/google/2020/03/08/the-unexpected-google-wide-domain-check-bypass/