Allow for modification of certain components
When using the URIBulider, it's plausible that one would want to use a
base URI and then modify components like the path and query string
without replacing them wholesale. Let's add two methods for that as well
as a convenience method to replace the `.finalize().unsplit()` dance.

This updates our docs to be a bit friendlier and adds release notes for
this change.

Closes #29