Match empty path without using line begin/end
This change allows correclty recognizing URIs with a scheme, but
no authority and an empty path (and possibly non-empty query or
fragment) as absolute.  The "about:" URI scheme is a scheme
that allows such syntax.

This uses a non-capturing group as it was unclear if any code
anywhere else was relying on capture group numbering from regexes
that include the empty path regex.